# openapi.json
{
  "components": {
    "schemas": {
...
      "ClickIn": {
        "properties": {
          "callback_data": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Callback Data",
            "type": "string"
          },
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "confirm_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Token"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_.:-]{1,64}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "chat_id",
          "message_id",
          "revision",
          "callback_data",
          "idempotency_key"
        ],
        "title": "ClickIn",
        "type": "object"
      },
...
      "CommandOut": {
        "properties": {
          "action_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "status",
          "reason"
        ],
        "title": "CommandOut",
        "type": "object"
      },
      "ConfirmRequired": {
        "properties": {
          "code": {
            "const": "confirm_required",
            "default": "confirm_required",
            "title": "Code",
            "type": "string"
          },
          "command_class": {
            "title": "Command Class",
            "type": "string"
          },
          "confirm_token": {
            "title": "Confirm Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "reason": {
            "enum": [
              "missing",
              "invalid",
              "expired"
            ],
            "title": "Reason",
            "type": "string"
          },
          "state_version": {
            "title": "State Version",
            "type": "integer"
          }
        },
        "required": [
          "reason",
          "confirm_token",
          "expires_at",
          "state_version",
          "command_class"
        ],
        "title": "ConfirmRequired",
        "type": "object"
      },
      "ConfirmRequiredOut": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ConfirmRequired"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ConfirmRequiredOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "description": "Ошибка со строковым кодом.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "ParamSpec": {
        "description": "Обязательный параметр сценария: `values` — у enum, `min`/`max` — у int, `pattern` — у\nstring (ECMA-262). Проверяет сервер: запуск с недопустимым значением — 422.",
        "properties": {
          "max": {
            "title": "Max",
            "type": "integer"
          },
          "min": {
            "title": "Min",
            "type": "integer"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "type": {
            "enum": [
              "enum",
              "int",
              "string"
            ],
            "title": "Type",
            "type": "string"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "type"
        ],
        "title": "ParamSpec",
        "type": "object"
      },
...
      "ScenarioInfo": {
        "properties": {
          "certified": {
            "title": "Certified",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "required": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ParamSpec"
            },
            "title": "Required",
            "type": "object"
          }
        },
        "required": [
          "name",
          "certified",
          "params",
          "required"
        ],
        "title": "ScenarioInfo",
        "type": "object"
      },
      "ScenarioRunAccepted": {
        "properties": {
          "scenario_run_id": {
            "title": "Scenario Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scenario_run_id",
          "status"
        ],
        "title": "ScenarioRunAccepted",
        "type": "object"
      },
...
      "ScenarioRunIn": {
        "properties": {
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_.:-]{1,64}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "ScenarioRunIn",
        "type": "object"
      },
...
      "SendIn": {
        "properties": {
          "confirm_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Token"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_.:-]{1,64}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "text": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "idempotency_key"
        ],
        "title": "SendIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "pyrobot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commands/click": {
      "post": {
        "operationId": "command_click_api_v1_commands_click_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClickIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            },
            "description": "Not finished yet; repeat with the same key"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "csrf token mismatch | forbidden | donate"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmRequiredOut"
                }
              }
            },
            "description": "risky command needs confirm_token"
          },
          "422": {
            "description": "invalid body or idempotency_key reused with other parameters"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "engine not started | store_failed"
          }
        },
        "summary": "Command Click",
        "tags": [
          "commands"
        ]
      }
    },
    "/api/v1/commands/send": {
      "post": {
        "operationId": "command_send_api_v1_commands_send_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            },
            "description": "Not finished yet; repeat with the same key"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "csrf token mismatch | forbidden | donate"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmRequiredOut"
                }
              }
            },
            "description": "risky command needs confirm_token"
          },
          "422": {
            "description": "invalid body or idempotency_key reused with other parameters"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "engine not started | store_failed"
          }
        },
        "summary": "Command Send",
        "tags": [
          "commands"
        ]
      }
    },
...
    "/api/v1/scenarios": {
      "get": {
        "description": "Сценарии, доступные ручному запуску; несертифицированные исполняются как simulate.",
        "operationId": "scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioInfo"
                  },
                  "title": "Response Scenarios Api V1 Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          }
        },
        "summary": "Scenarios",
        "tags": [
          "commands"
        ]
      }
    },
    "/api/v1/scenarios/{name}/run": {
      "post": {
        "operationId": "scenario_run_api_v1_scenarios__name__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRunAccepted"
                }
              }
            },
            "description": "Key already used: existing run"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "csrf token mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unknown scenario | scenario run not found"
          },
          "422": {
            "description": "invalid body, params contradicting fixed scenario params, missing or invalid required params, or idempotency_key reused with other parameters"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "engine not started | planner not started"
          }
        },
        "summary": "Scenario Run",
        "tags": [
          "commands"
        ]
      }
    },
...
  }
}